move evaluation form into sidebar
